requirements.txt
langchain_community   
langchainhub 
chromadb 
langchain 
langgraph  
python-dotenv
google-generativeai
langchain-google-genai
langchain-text-splitters
langchain_huggingface
pytz
flask
